+++ app/uk/gov/hmrc/gform/submission/SubmissionRef.scala
@@ -16,11 +16,52 @@
 
 package uk.gov.hmrc.gform.submission
 
+import java.math.BigInteger
+import java.security.MessageDigest
+
 import play.api.libs.json.OFormat
+import uk.gov.hmrc.gform.submission.SubmissionRef.{ calculate, calculateCheckCharacter, verify }
 import uk.gov.hmrc.gform.sharedmodel.ValueClassFormat
+import uk.gov.hmrc.gform.sharedmodel.form.EnvelopeId
+
+import scala.math.pow
 
 case class SubmissionRef(value: String)
 
 object SubmissionRef {
+  val radix = 36
+  val digits = 11
+  val comb: Stream[Int] = Stream.continually(List(1, 3)).flatten
+
   val oformat: OFormat[SubmissionRef] = ValueClassFormat.oformat("submissionRef", SubmissionRef.apply, _.value)
+
+  def getSubmissionReference(envelopeId: EnvelopeId): String =
+    if (!envelopeId.value.isEmpty) {
+      // As 36^11 (number of combinations of 11 base 36 digits) < 2^63 (number of combinations of 63 base 2 digits) we can get full significance from this digest.
+      val digest = MessageDigest.getInstance("SHA-256").digest(envelopeId.value.getBytes()).take(8)
+      val initialValue = new BigInteger(digest).abs()
+      val unformattedString = calculate(initialValue, radix, digits, comb)
+      unformattedString.grouped(4).mkString("-").toUpperCase
+    } else { "" }
+
+  def calculate(value: BigInteger, radix: Int, digits: Int, comb: Stream[Int]): String = {
+    val modulus: BigInteger = BigInteger.valueOf(pow(radix, digits).toLong)
+    val derivedDigits = (value.mod(modulus) add modulus).toString(radix).takeRight(digits)
+    val checkCharacter = BigInteger.valueOf(calculateCheckCharacter(derivedDigits, radix, comb)).toString(radix)
+    checkCharacter + derivedDigits
+  }
+
+  def calculateCheckCharacter(digits: String, radix: Int, comb: Stream[Int]): Int = {
+    val stringToInts = digits.toCharArray.map(i => Integer.parseInt(i.toString, radix))
+    stringToInts.zip(comb).map(i => i._1 * i._2).sum % radix
+  }
+
+  def verifyCheckChar(reference: String): Boolean =
+    "^([A-Z0-9]{4})-([A-Z0-9]{4})-([A-Z0-9]{4})$".r.findFirstMatchIn(reference) match {
+      case Some(a) => verify(s"${a.group(1)}${a.group(2)}${a.group(3)}", radix, comb)
+      case _       => false
+    }
+
+  def verify(reference: String, radix: Int, comb: Stream[Int]): Boolean =
+    calculateCheckCharacter(reference.tail, radix, comb) == Integer.parseInt(reference.head.toString, radix)
 }
